+++ src/platform_impl/macos/window_delegate.rs
@@ -450,8 +450,7 @@ extern "C" fn window_will_enter_fullscreen(this: &Object, _: Sel, _: id) {
                 // Otherwise, we must've reached fullscreen by the user clicking
                 // on the green fullscreen button. Update state!
                 None => {
-                    let current_monitor = window.current_monitor_inner();
-                    shared_state.fullscreen = Some(Fullscreen::Borderless(current_monitor))
+                    shared_state.fullscreen = Some(Fullscreen::Borderless(window.current_monitor()))
                 }
             }
             shared_state.in_fullscreen_transition = true;