+++ src/platform_impl/windows/event_loop.rs
@@ -853,7 +853,8 @@ unsafe extern "system" fn public_window_callback<T: 'static>(
                     winuser::MonitorFromRect(&new_rect, winuser::MONITOR_DEFAULTTONULL);
                 match fullscreen {
                     Fullscreen::Borderless(ref mut fullscreen_monitor) => {
-                        if new_monitor != fullscreen_monitor.inner.hmonitor()
+                        // FIXME This should be properly implemented.
+                        if new_monitor != fullscreen_monitor.as_ref().unwrap().inner.hmonitor()
                             && new_monitor != ptr::null_mut()
                         {
                             if let Ok(new_monitor_info) = monitor::get_monitor_info(new_monitor) {
@@ -863,9 +864,9 @@ unsafe extern "system" fn public_window_callback<T: 'static>(
                                 window_pos.cx = new_monitor_rect.right - new_monitor_rect.left;
                                 window_pos.cy = new_monitor_rect.bottom - new_monitor_rect.top;
                             }
-                            *fullscreen_monitor = crate::monitor::MonitorHandle {
+                            *fullscreen_monitor = Some(crate::monitor::MonitorHandle {
                                 inner: MonitorHandle::new(new_monitor),
-                            };
+                            });
                         }
                     }
                     Fullscreen::Exclusive(ref video_mode) => {